backend_la/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
drf-yasg==1.21.7
ecdsa==0.18.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-jose==3.3.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
resend==0.7.2
rsa==4.9
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.0
whitenoise==6.6.0
